[tilegx] Fix LCodeGen::PrepareKeyedOperand bug.
  LCodeGen::PrepareKeyedOperand was wrong when handling Smi (v8 boxed
  number format).

    Also fixed MacroAssembler::target_address_at bug. The decoded bits
  needs sign extension. Although our vaddress is always <=42bit, but for v8,
  there are some bogus address, for example 0xfffffffffffe0000 generated by
  ExternalReference::new_space_mask.

+++ src/tilegx/assembler-tilegx.cc
@@ -944,7 +944,7 @@ Address Assembler::target_address_at(Address pc) {
     // Assemble the 48 bit value.
     // FIXME: currently, we always assume they are encoded in X1 slot.
     return reinterpret_cast<Address>(
-        ((long)get_Imm16_X1(instr1) << 32) | (get_Imm16_X1(instr2) << 16) | get_Imm16_X1(instr3));
+        ((long)((short)get_Imm16_X1(instr1)) << 32) | (get_Imm16_X1(instr2) << 16) | get_Imm16_X1(instr3));
   }
 
   // We should never get here, force a bad address if we do.
